# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Batch:
    properties:
      analysis_result: {}
      batch_id:
        type: string
      created_at:
        type: string
      expected_count:
        type: integer
      received_count:
        type: integer
      results:
        additionalProperties: true
        type: object
      status:
        $ref: '#/definitions/models.BatchStatus'
      user_id:
        type: string
    type: object
  models.BatchStatus:
    enum:
    - processing
    - complete
    type: string
    x-enum-varnames:
    - BatchStatusProcessing
    - BatchStatusComplete
...
  models.ScanStatus:
    enum:
    - pending
    - running
    - stopped
    - failed
    - completed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusStopped
    - StatusFailed
    - StatusCompleted
  models.ScanTaskResponse:
    properties:
      batch_id:
        type: string
      error:
        type: string
      progress:
        type: integer
      result:
        items:
          additionalProperties: true
          type: object
        type: array
      started_at:
        description: RFC3339 format
        type: string
      status:
        $ref: '#/definitions/models.ScanStatus'
      target:
        type: string
      task_id:
        type: string
      updated_at:
        description: RFC3339 format
        type: string
      user_id:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: API Support
  description: Security Scanning API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Napscan API
  version: "1.0"
paths:
...
  /batch/create:
    post:
      description: Generate a unique batch ID for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              batch_id:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Batch ID
      tags:
      - Batch
  /batch/list:
    get:
      description: Get a list of all batches created by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Batch'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get User Batches
      tags:
      - Batch
  /ffuf/scan:
    post:
      consumes:
      - application/json
      description: Run directory fuzzing using FFUF
      parameters:
      - description: Target URL
        in: body
        name: target
        required: true
        schema:
          properties:
            batch_id:
              type: string
            target:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Start FFUF Scan
      tags:
      - FFUF
...
  /ffuf/scan/async:
    post:
      consumes:
      - application/json
      description: Start a FFUF scan asynchronously. Returns task_id immediately.
      parameters:
      - description: Scan Request
        in: body
        name: request
        required: true
        schema:
          properties:
            batch_id:
              type: string
            target:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Start FFUF Scan (Async)
      tags:
      - FFUF
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"